Debugging tools: Chunk Slice Viewer
Added a debug level variable to the chunk manager
If debug level is greater than 1 a new ChunkDataViewer script is added to all chunk objects which allows slices of the density map to be viewed in the inspector. These slices indicate if the chunk is editable or not with their color. Green = editable, Red = not editable.
These tools will help fix the errors in transitioning from editable to non editable density maps.