Continentalness & Erosion
Added Continentalness and Erosion to noise generation shader to control the squashing factor of the base 3D noise. I think the scale and mins/maxs are good for now but could always be tuned later.

Need to add spline points to the continentaless and erosion values to have more control over terrain shape.

I had a variable called continentalness before but it was not actual continentalness.