Refining Ledges and Downhill + Better Spawning
Added a check to see if the player is moving the expected amount due to gravity. If a certain percentage is not met, acceleration due to gravity is frozen until the threshold is met.

If the player becomes un-grounded but is still withing 3x ground check distance, an initial burst of downwards speed is given to better follow the slope.